requirements.txt
requests
feedparser
python-dotenv
python-telegram-bot
openai
sqlalchemy
schedule
